EU-115 refactor authentication
