+++ apps/authorization/handlers.py
@@ -1,24 +1,14 @@
-from apps.common.utils import (create_access_token, create_refresh_token, verify_password)
+from apps.authorization.services import (create_access_token, create_refresh_token, verify_password)
 from apps.authorization.models import Role
-from apps.users.handlers import get_current_user
 from apps.common.exceptions import BackendException
-from apps.authorization.schemas import AvailableRolesOut
-from apps.common.schemas import JSENDOutSchema
 from apps.users.models import User
-from apps.users.schemas import AuthOut
 from apps.common.db import database
 
-from typing import List
-from fastapi import Depends, APIRouter
+from fastapi import Depends
 from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy import select
 
 
-authorization_router = APIRouter()
-
-
-@authorization_router.post('/login', summary="Створення доступу та оновлення токена користувача", response_model=AuthOut,
-                           tags=["Authorization"])
 async def login(form_data: OAuth2PasswordRequestForm = Depends()):
     query = select(User).where(User.login == form_data.username)
     user = await database.fetch_one(query)
@@ -41,10 +31,6 @@ async def login(form_data: OAuth2PasswordRequestForm = Depends()):
     }
 
 
-@authorization_router.get("/roles/", response_model=JSENDOutSchema[List[AvailableRolesOut]], tags=["SuperAdmin dashboard"])
-async def available_roles(user=Depends(get_current_user)):
+async def available_roles():
     query = select(Role)
-    return {
-        "data": await database.fetch_all(query),
-        "message": "Got roles"
-    }
+    return await database.fetch_all(query)
